Reoganize POS Presentation layer
