+++ Android/app/src/main/java/live/ditto/pos/MainActivity.kt
@@ -3,16 +3,8 @@ package live.ditto.pos
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Surface
-import androidx.compose.ui.Modifier
-import androidx.navigation.compose.rememberNavController
 import dagger.hilt.android.AndroidEntryPoint
-import live.ditto.pos.core.presentation.composables.PosKdsNavigationBar
-import live.ditto.pos.core.presentation.navigation.BottomNavItem
-import live.ditto.pos.core.presentation.navigation.PosKdsNavHost
-import live.ditto.pos.ui.theme.DittoPoSKDSDemoTheme
+import live.ditto.pos.core.presentation.screens.PosKdsApp
 
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
@@ -20,30 +12,7 @@ class MainActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
 
         setContent {
-            val bottomNavItems = listOf(
-                BottomNavItem.PointOfSale,
-                BottomNavItem.KitchenDisplay
-            )
-            val navHostController = rememberNavController()
-
-            DittoPoSKDSDemoTheme {
-                Surface {
-                    Scaffold(
-                        bottomBar = {
-                            PosKdsNavigationBar(
-                                bottomNavItems = bottomNavItems,
-                            ) {
-                                navHostController.navigate(route = it.route)
-                            }
-                        },
-                        content = {
-                            Surface(modifier = Modifier.padding(it)) {
-                                PosKdsNavHost(navHostController = navHostController)
-                            }
-                        }
-                    )
-                }
-            }
+            PosKdsApp()
         }
     }
 }
